feat: Add "Read more" button to show/hide text
Added a "Read more" button and functionality to the Welcome message.
This allows the reader to show/hide part of the message and preserves
a nice proportional layout. Functionality of the "Read more"/"Read less"
button was implemented with Knockout bindings in HTML, supporting
observables in the viewModel, and CSS styling.